--- src/test/java/wayc/backend/unit/common/ItemControllerTest.java
@@ -1,7 +1,0 @@
-package wayc.backend.unit.common;
-
-import wayc.backend.unit.ControllerTest;
-
-public class ItemControllerTest extends ControllerTest {
-
-}
