+++ src/test/java/wayc/backend/unit/presentation/CommonExceptionTest.java
@@ -1,12 +1,10 @@
-package wayc.backend.unit;
+package wayc.backend.unit.presentation;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.MediaType;
-import org.springframework.restdocs.RestDocumentationContextProvider;
 import org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders;
-import org.springframework.web.context.WebApplicationContext;
+import wayc.backend.unit.presentation.ControllerTest;
 
 import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
 import static org.springframework.restdocs.payload.JsonFieldType.STRING;
