+++ src/test/java/wayc/backend/unit/presentation/ControllerTest.java
@@ -1,4 +1,4 @@
-package wayc.backend.unit;
+package wayc.backend.unit.presentation;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
